+++ packages/react-dev-overlay/src/internal/StackFrame.ts
