+++ packages/react-dev-overlay/src/internal/noop-template.ts
@@ -1,0 +1,7 @@
+export function noop(strings, ...keys) {
+  const lastIndex = strings.length - 1
+  return (
+    strings.slice(0, lastIndex).reduce((p, s, i) => p + s + keys[i], '') +
+    strings[lastIndex]
+  )
+}
