+++ packages/react-dev-overlay/src/internal/stackFrameParser.ts
